Adds a "links" directory.
Just like "drafts", this is for content that is like a post
but should not be linked to from the main pages.

Unlike a draft, this is for information that we want to
disitribute in a more targetted way than a post.  They aren't
necessarily sensitive - the content is going to our public
Github after all -  but adding links to them from people and
project pages might be a bit noisy / off-topic for the target
audience of those pages.