Added more information regarding textInput Event relationship.
textInput_input Event is not triggered when the textInput_setValue Event is
fired. This meant the knot can be tied in a single Event containing all of the
update logic, and the rest of the widget is free to use it without fear of
looping the textInput value out of space and time.