Improved Project(s), People(s), and Location pages.
Added carousel of the images to the location page. Improved
the layout of the assorted methods of locating the office.

Fixed up the heading and layout of the project pages. Added
class to allow for slight massaging of H2 headers into a
more manageable size when displaying items in a list.

Improved the layout of the people pages a bit to be less lout
, err loud. Generally reducing heading sizes etc.

Tweaked the location page to be a full HTML page now instead
of the markdown format.

Added carousel init javascript