Added 'links' to Project template.
The three most recent links added to the `/links` folder that have a related `project:` attribute will appear on that projects page. This should make it easier to add information about an upcoming course without having to repeat the requirements or outcomes each time.

Removed some older links from the `/links` folder.

Updated the `release.nix` and `generator.nix` to accept a `compiler` string argument to let the user specify which `ghc` they want to use to build the generator.

Pinned the version of `nixpkgs` that we're using to build the blog. By default the blog will build with `ghc802` and nixpkgs version `17.09`. These can be overridden.